refactored AnimatedProgressBar
